fix aws sample applications
