+++ Aspire.sln
@@ -158,13 +158,13 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Aspire.AWS", "src\Component
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "aws", "aws", "{4FE4E999-1D3B-466F-802E-0389EF07EC8E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AwsAppHost.AppHost", "samples\aws\AwsAppHost\AwsAppHost.AppHost\AwsAppHost.AppHost.csproj", "{CF63B177-463A-42F4-BC34-6AD05702E754}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AwsAppHost.ServiceDefaults", "samples\aws\AwsAppHost\AwsAppHost.ServiceDefaults\AwsAppHost.ServiceDefaults.csproj", "{E320453A-DF78-4676-94DF-E5568FC4A548}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "UserService", "samples\aws\AwsAppHost\UserService\UserService.csproj", "{202D78B2-CC0D-4F18-A7E1-A20F12D686A7}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Aspire.Hosting.AWS.Provisioning", "src\Aspire.Hosting.AWS.Provisioning\Aspire.Hosting.AWS.Provisioning.csproj", "{A2A42422-6FC1-4B3C-A495-DAC86829A381}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Aws.AppHost", "samples\aws\Aws.AppHost\Aws.AppHost.csproj", "{F851D31E-1D3A-40DA-B11D-1138179B866D}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Aws.ServiceDefaults", "samples\aws\Aws.ServiceDefaults\Aws.ServiceDefaults.csproj", "{8611076F-CA70-4980-808F-BB9AF49BFCFB}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Aws.UserService", "samples\aws\Aws.UserService\Aws.UserService.csproj", "{9A7CADAA-1B13-40FE-90FA-2B34491C9B55}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -428,22 +428,22 @@ Global
 		{B2B9CCC6-3DC5-45AA-BBC0-A4258286828B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B2B9CCC6-3DC5-45AA-BBC0-A4258286828B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B2B9CCC6-3DC5-45AA-BBC0-A4258286828B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{CF63B177-463A-42F4-BC34-6AD05702E754}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CF63B177-463A-42F4-BC34-6AD05702E754}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CF63B177-463A-42F4-BC34-6AD05702E754}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CF63B177-463A-42F4-BC34-6AD05702E754}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E320453A-DF78-4676-94DF-E5568FC4A548}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E320453A-DF78-4676-94DF-E5568FC4A548}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E320453A-DF78-4676-94DF-E5568FC4A548}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E320453A-DF78-4676-94DF-E5568FC4A548}.Release|Any CPU.Build.0 = Release|Any CPU
-		{202D78B2-CC0D-4F18-A7E1-A20F12D686A7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{202D78B2-CC0D-4F18-A7E1-A20F12D686A7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{202D78B2-CC0D-4F18-A7E1-A20F12D686A7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{202D78B2-CC0D-4F18-A7E1-A20F12D686A7}.Release|Any CPU.Build.0 = Release|Any CPU
 		{A2A42422-6FC1-4B3C-A495-DAC86829A381}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A2A42422-6FC1-4B3C-A495-DAC86829A381}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A2A42422-6FC1-4B3C-A495-DAC86829A381}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A2A42422-6FC1-4B3C-A495-DAC86829A381}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F851D31E-1D3A-40DA-B11D-1138179B866D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F851D31E-1D3A-40DA-B11D-1138179B866D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F851D31E-1D3A-40DA-B11D-1138179B866D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F851D31E-1D3A-40DA-B11D-1138179B866D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8611076F-CA70-4980-808F-BB9AF49BFCFB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8611076F-CA70-4980-808F-BB9AF49BFCFB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8611076F-CA70-4980-808F-BB9AF49BFCFB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8611076F-CA70-4980-808F-BB9AF49BFCFB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9A7CADAA-1B13-40FE-90FA-2B34491C9B55}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9A7CADAA-1B13-40FE-90FA-2B34491C9B55}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9A7CADAA-1B13-40FE-90FA-2B34491C9B55}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9A7CADAA-1B13-40FE-90FA-2B34491C9B55}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -518,10 +518,10 @@ Global
 		{7738E898-5D00-45C2-A498-EA7C56E31A27} = {B80354C7-BE58-43F6-8928-9F3A74AB7F47}
 		{B2B9CCC6-3DC5-45AA-BBC0-A4258286828B} = {27381127-6C45-4B4C-8F18-41FF48DFE4B2}
 		{4FE4E999-1D3B-466F-802E-0389EF07EC8E} = {D173887B-AF42-4576-B9C1-96B9E9B3D9C0}
-		{CF63B177-463A-42F4-BC34-6AD05702E754} = {4FE4E999-1D3B-466F-802E-0389EF07EC8E}
-		{E320453A-DF78-4676-94DF-E5568FC4A548} = {4FE4E999-1D3B-466F-802E-0389EF07EC8E}
-		{202D78B2-CC0D-4F18-A7E1-A20F12D686A7} = {4FE4E999-1D3B-466F-802E-0389EF07EC8E}
 		{A2A42422-6FC1-4B3C-A495-DAC86829A381} = {B80354C7-BE58-43F6-8928-9F3A74AB7F47}
+		{F851D31E-1D3A-40DA-B11D-1138179B866D} = {4FE4E999-1D3B-466F-802E-0389EF07EC8E}
+		{8611076F-CA70-4980-808F-BB9AF49BFCFB} = {4FE4E999-1D3B-466F-802E-0389EF07EC8E}
+		{9A7CADAA-1B13-40FE-90FA-2B34491C9B55} = {4FE4E999-1D3B-466F-802E-0389EF07EC8E}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {6DCEDFEC-988E-4CB3-B45B-191EB5086E0C}
