+++ samples/aws/Aws.AppHost/Aws.AppHost.csproj
@@ -1,0 +1,19 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net8.0</TargetFramework>
+    <UserSecretsId>2bf0840a-0dfc-82aa-1586-def9b3b31da0</UserSecretsId>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <IsAspireHost>true</IsAspireHost>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\..\src\Aspire.Hosting\Aspire.Hosting.csproj" />
+    <ProjectReference Include="..\..\..\src\Aspire.Hosting.Aws\Aspire.Hosting.Aws.csproj" />
+    <ProjectReference Include="..\..\..\src\Aspire.Hosting.Aws.Provisioning\Aspire.Hosting.Aws.Provisioning.csproj" />
+
+    <ProjectReference Include="..\Aws.UserService\Aws.UserService.csproj" />
+  </ItemGroup>
+
+</Project>
