+++ samples/aws/Aws.AppHost/appsettings.json
@@ -1,0 +1,22 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning",
+      "Aspire.Hosting.Dcp": "Warning"
+    }
+  },
+  "LocalStack": {
+    "UseLocalStack": true
+  },
+  "AWS": {
+    "Profile": "personal",
+    "Region": "eu-central-1",
+    "S3": {
+      "ProfilePicturesBucket": {
+        "BucketName": "aspire-test-profile-pic-bkt",
+        "Region": "eu-central-1"
+      }
+    }
+  }
+}
