+++ samples/aws/Aws.ServiceDefaults/Extensions.cs
