+++ samples/aws/Aws.UserService/appsettings.Development.json
