--- samples/aws/AwsAppHost/AwsAppHost.AppHost/AwsAppHost.AppHost.csproj
@@ -1,16 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net8.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <IsAspireHost>true</IsAspireHost>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\..\src\Aspire.Hosting.AWS\Aspire.Hosting.AWS.csproj" />
-    <ProjectReference Include="..\..\..\..\src\Aspire.Hosting\Aspire.Hosting.csproj" />
-  </ItemGroup>
-
-</Project>
