--- samples/aws/AwsAppHost/AwsAppHost.AppHost/Properties/launchSettings.json
@@ -1,28 +1,0 @@
-{
-  "$schema": "http://json.schemastore.org/launchsettings.json",
-  "profiles": {
-    "http": {
-      "commandName": "Project",
-      "dotnetRunMessages": true,
-      "launchBrowser": true,
-      "applicationUrl": "http://localhost:15180",
-      "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development",
-        "DOTNET_ENVIRONMENT": "Development",
-        "DOTNET_DASHBOARD_OTLP_ENDPOINT_URL": "http://localhost:16031"
-      }
-    }
-  },
-  "generate-manifest": {
-    "commandName": "Project",
-    "launchBrowser": true,
-    "dotnetRunMessages": true,
-    "commandLineArgs": "--publisher manifest --output-path aspire-manifest.json",
-    "applicationUrl": "http://localhost:15888",
-    "environmentVariables": {
-      "ASPNETCORE_ENVIRONMENT": "Development",
-      "DOTNET_ENVIRONMENT": "Development",
-      "DOTNET_DASHBOARD_OTLP_ENDPOINT_URL": "http://localhost:16031"
-    }
-  }
-}
