--- samples/aws/AwsAppHost/AwsAppHost.AppHost/aspire-manifest.json
@@ -1,3 +1,0 @@
-{
-  "resources": {}
-}
