--- samples/aws/AwsAppHost/UserService/Dockerfile
@@ -1,30 +1,0 @@
-#See https://aka.ms/customizecontainer to learn how to customize your debug container and how Visual Studio uses this Dockerfile to build your images for faster debugging.
-
-FROM mcr.microsoft.com/dotnet/aspnet:8.0 AS base
-USER app
-WORKDIR /app
-EXPOSE 8080
-EXPOSE 8081
-
-FROM mcr.microsoft.com/dotnet/sdk:8.0 AS build
-ARG BUILD_CONFIGURATION=Release
-WORKDIR /src
-COPY ["Directory.Packages.props", "."]
-COPY ["samples/Directory.Build.props", "samples/"]
-COPY ["Directory.Build.props", "."]
-COPY ["Directory.Build.targets", "."]
-COPY ["NuGet.config", "."]
-COPY ["samples/aws/AwsAppHost/UserService/UserService.csproj", "samples/aws/AwsAppHost/UserService/"]
-RUN dotnet restore "./samples/aws/AwsAppHost/UserService/./UserService.csproj"
-COPY . .
-WORKDIR "/src/samples/aws/AwsAppHost/UserService"
-RUN dotnet build "./UserService.csproj" -c $BUILD_CONFIGURATION -o /app/build
-
-FROM build AS publish
-ARG BUILD_CONFIGURATION=Release
-RUN dotnet publish "./UserService.csproj" -c $BUILD_CONFIGURATION -o /app/publish /p:UseAppHost=false
-
-FROM base AS final
-WORKDIR /app
-COPY --from=publish /app/publish .
-ENTRYPOINT ["dotnet", "UserService.dll"]
